Implement Admin Panel into Main Menu - Attempt #7
